// apps/mobile/http/generated/models/PutOrganizationsSlugProductsId.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type PutOrganizationsSlugProductsIdPathParams = {
  /**
   * @type string
   */
  slug: string;
  /**
   * @type string, uuid
   */
  id: string;
};

export const putOrganizationsSlugProductsId204Enum = {} as const;

export type PutOrganizationsSlugProductsId204EnumKey =
  (typeof putOrganizationsSlugProductsId204Enum)[keyof typeof putOrganizationsSlugProductsId204Enum];

/**
 * @description Default Response
 */
export type PutOrganizationsSlugProductsId204 =
  PutOrganizationsSlugProductsId204EnumKey | null;

export type PutOrganizationsSlugProductsIdMutationRequest = {
  /**
   * @minLength 1
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  description: string | null;
  /**
   * @type string, uuid
   */
  parentId: string | null;
  /**
   * @type boolean
   */
  isActive: boolean;
  /**
   * @minLength 0
   * @maxLength 9007199254740991
   * @type integer
   */
  sortOrder: number;
};

export type PutOrganizationsSlugProductsIdMutationResponse =
  PutOrganizationsSlugProductsId204;

export type PutOrganizationsSlugProductsIdMutation = {
  Response: PutOrganizationsSlugProductsId204;
  Request: PutOrganizationsSlugProductsIdMutationRequest;
  PathParams: PutOrganizationsSlugProductsIdPathParams;
  Errors: any;
};